Added in background music, some cleanup.
Removed unnecessary files.
Music is currently located in main. They should probably be in 'stage'
Reference to audio is currently stored on the html file. I couldn't find
a better way to do this.
Added in Howler.js to support the sound
Added a noscript tag to the html file.
Added "credits" to the html file.