Music changes in Main
Added notes about the background music.
Corrected an error regarding BGM.castle. It should now loop correctly
Sent some more magic numbers in main to the shadow realm.
Experimented with the sky scrolling too. It scrolls 1/10th as quickly as
the trees.
Images of the sky at the start of the level and end of the level are
included.