Revert "complete animating archer and arrow, need to handle collision"
This reverts commit db820ef4837340e4b52f4983416e242105237a48.

I'm reverting this commit for several reasons.
1 - It looks like master was committed to directly. No one should be
doing this unless the change is very small.

2 - I am Immediately suspicious when 8 files are changed with 163
additions, 296 deletions, and a 1 line commit log.

3 - Many of the files changed seem to have little to do with the commit
log.

4 - Nearly everything in the agents folder calls draw with 2 parameters
when previously they did not and drew perfectly fine. What was the
reasoning behind this?

5 - Agents now receive a bunch of information most of them don't
require, and that all of them are able to access through their own game
engine reference. Same question: Why?